+++ Example/web3swiftExample/Pods/web3swift/web3swift/Web3/Classes/Web3+Utils.swift
@@ -10,7 +10,7 @@ import Foundation
 import BigInt
 import CryptoSwift
 
-public typealias Web3Utils = Web3.Utils
+public typealias Web3Utils = Web3Utils
 
 extension Web3 {
     public struct Utils {
@@ -19,11 +19,11 @@ extension Web3 {
     }
 }
 
-extension Web3.Utils {
+extension Web3Utils {
     public static func calcualteContractAddress(from: EthereumAddress, nonce: BigUInt) -> EthereumAddress? {
         guard let normalizedAddress = from.addressData.setLengthLeft(32) else {return nil}
         guard let data = RLP.encode([normalizedAddress, nonce] as [Any]) else {return nil}
-        guard let contractAddressData = Web3.Utils.sha3(data)?[12..<32] else {return nil}
+        guard let contractAddressData = Web3Utils.sha3(data)?[12..<32] else {return nil}
         guard let contractAddress = EthereumAddress(Data(contractAddressData)) else {return nil}
         return contractAddress
     }
@@ -63,7 +63,7 @@ extension Web3.Utils {
     public static var erc20ABI = "[{\"constant\":true,\"inputs\":[],\"name\":\"name\",\"outputs\":[{\"name\":\"\",\"type\":\"string\"}],\"payable\":false,\"type\":\"function\"},{\"constant\":false,\"inputs\":[{\"name\":\"_spender\",\"type\":\"address\"},{\"name\":\"_value\",\"type\":\"uint256\"}],\"name\":\"approve\",\"outputs\":[{\"name\":\"success\",\"type\":\"bool\"}],\"payable\":false,\"type\":\"function\"},{\"constant\":true,\"inputs\":[],\"name\":\"totalSupply\",\"outputs\":[{\"name\":\"\",\"type\":\"uint256\"}],\"payable\":false,\"type\":\"function\"},{\"constant\":false,\"inputs\":[{\"name\":\"_from\",\"type\":\"address\"},{\"name\":\"_to\",\"type\":\"address\"},{\"name\":\"_value\",\"type\":\"uint256\"}],\"name\":\"transferFrom\",\"outputs\":[{\"name\":\"success\",\"type\":\"bool\"}],\"payable\":false,\"type\":\"function\"},{\"constant\":true,\"inputs\":[],\"name\":\"decimals\",\"outputs\":[{\"name\":\"\",\"type\":\"uint8\"}],\"payable\":false,\"type\":\"function\"},{\"constant\":true,\"inputs\":[],\"name\":\"version\",\"outputs\":[{\"name\":\"\",\"type\":\"string\"}],\"payable\":false,\"type\":\"function\"},{\"constant\":true,\"inputs\":[{\"name\":\"_owner\",\"type\":\"address\"}],\"name\":\"balanceOf\",\"outputs\":[{\"name\":\"balance\",\"type\":\"uint256\"}],\"payable\":false,\"type\":\"function\"},{\"constant\":true,\"inputs\":[],\"name\":\"symbol\",\"outputs\":[{\"name\":\"\",\"type\":\"string\"}],\"payable\":false,\"type\":\"function\"},{\"constant\":false,\"inputs\":[{\"name\":\"_to\",\"type\":\"address\"},{\"name\":\"_value\",\"type\":\"uint256\"}],\"name\":\"transfer\",\"outputs\":[{\"name\":\"success\",\"type\":\"bool\"}],\"payable\":false,\"type\":\"function\"},{\"constant\":false,\"inputs\":[{\"name\":\"_spender\",\"type\":\"address\"},{\"name\":\"_value\",\"type\":\"uint256\"},{\"name\":\"_extraData\",\"type\":\"bytes\"}],\"name\":\"approveAndCall\",\"outputs\":[{\"name\":\"success\",\"type\":\"bool\"}],\"payable\":false,\"type\":\"function\"},{\"constant\":true,\"inputs\":[{\"name\":\"_owner\",\"type\":\"address\"},{\"name\":\"_spender\",\"type\":\"address\"}],\"name\":\"allowance\",\"outputs\":[{\"name\":\"remaining\",\"type\":\"uint256\"}],\"payable\":false,\"type\":\"function\"},{\"inputs\":[{\"name\":\"_initialAmount\",\"type\":\"uint256\"},{\"name\":\"_tokenName\",\"type\":\"string\"},{\"name\":\"_decimalUnits\",\"type\":\"uint8\"},{\"name\":\"_tokenSymbol\",\"type\":\"string\"}],\"type\":\"constructor\"},{\"payable\":false,\"type\":\"fallback\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":true,\"name\":\"_from\",\"type\":\"address\"},{\"indexed\":true,\"name\":\"_to\",\"type\":\"address\"},{\"indexed\":false,\"name\":\"_value\",\"type\":\"uint256\"}],\"name\":\"Transfer\",\"type\":\"event\"},{\"anonymous\":false,\"inputs\":[{\"indexed\":true,\"name\":\"_owner\",\"type\":\"address\"},{\"indexed\":true,\"name\":\"_spender\",\"type\":\"address\"},{\"indexed\":false,\"name\":\"_value\",\"type\":\"uint256\"}],\"name\":\"Approval\",\"type\":\"event\"},]"
 }
 
-extension Web3.Utils {
+extension Web3Utils {
     
     public static func privateToPublic(_ privateKey: Data, compressed: Bool = false) -> Data? {
         guard let publicKey = SECP256K1.privateToPublic(privateKey:  privateKey, compressed: compressed) else {return nil}
@@ -91,13 +91,13 @@ extension Web3.Utils {
     }
     
     public static func publicToAddress(_ publicKey: Data) -> EthereumAddress? {
-        guard let addressData = Web3.Utils.publicToAddressData(publicKey) else {return nil}
+        guard let addressData = Web3Utils.publicToAddressData(publicKey) else {return nil}
         let address = addressData.toHexString().addHexPrefix().lowercased()
         return EthereumAddress(address)
     }
     
     public static func publicToAddressString(_ publicKey: Data) -> String? {
-        guard let addressData = Web3.Utils.publicToAddressData(publicKey) else {return nil}
+        guard let addressData = Web3Utils.publicToAddressData(publicKey) else {return nil}
         let address = addressData.toHexString().addHexPrefix().lowercased()
         return address
     }
@@ -121,7 +121,7 @@ extension Web3.Utils {
         return hash
     }
     
-    public static func parseToBigUInt(_ amount: String, units: Web3.Utils.Units = .eth) -> BigUInt? {
+    public static func parseToBigUInt(_ amount: String, units: Web3Utils.Units = .eth) -> BigUInt? {
         let unitDecimals = units.decimals
         return parseToBigUInt(amount, decimals: unitDecimals)
     }
@@ -143,7 +143,7 @@ extension Web3.Utils {
         return mainPart
     }
     
-    public static func formatToEthereumUnits(_ bigNumber: BigInt, toUnits: Web3.Utils.Units = .eth, decimals: Int = 4, decimalSeparator: String = ".") -> String? {
+    public static func formatToEthereumUnits(_ bigNumber: BigInt, toUnits: Web3Utils.Units = .eth, decimals: Int = 4, decimalSeparator: String = ".") -> String? {
         let magnitude = bigNumber.magnitude
         guard let formatted = formatToEthereumUnits(magnitude, toUnits: toUnits, decimals: decimals, decimalSeparator: decimalSeparator) else {return nil}
         switch bigNumber.sign {
@@ -165,7 +165,7 @@ extension Web3.Utils {
         }
     }
     
-    public static func formatToEthereumUnits(_ bigNumber: BigUInt, toUnits: Web3.Utils.Units = .eth, decimals: Int = 4, decimalSeparator: String = ".", fallbackToScientific: Bool = false) -> String? {
+    public static func formatToEthereumUnits(_ bigNumber: BigUInt, toUnits: Web3Utils.Units = .eth, decimals: Int = 4, decimalSeparator: String = ".", fallbackToScientific: Bool = false) -> String? {
         return formatToPrecision(bigNumber, numberDecimals: toUnits.decimals, formattingDecimals: decimals, decimalSeparator: decimalSeparator, fallbackToScientific: fallbackToScientific);
     }
     
@@ -208,7 +208,7 @@ extension Web3.Utils {
     static public func personalECRecover(_ personalMessage: String, signature: String) -> EthereumAddress? {
         guard let data = Data.fromHex(personalMessage) else {return nil}
         guard let sig = Data.fromHex(signature) else {return nil}
-        return Web3.Utils.personalECRecover(data, signature:sig)
+        return Web3Utils.personalECRecover(data, signature:sig)
     }
     
     static public func personalECRecover(_ personalMessage: Data, signature: Data) -> EthereumAddress? {
@@ -217,9 +217,9 @@ extension Web3.Utils {
         let sData = signature[32..<64].bytes
         let vData = signature[64]
         guard let signatureData = SECP256K1.marshalSignature(v: vData, r: rData, s: sData) else {return nil}
-        guard let hash = Web3.Utils.hashPersonalMessage(personalMessage) else {return nil}
+        guard let hash = Web3Utils.hashPersonalMessage(personalMessage) else {return nil}
         guard let publicKey = SECP256K1.recoverPublicKey(hash: hash, signature: signatureData) else {return nil}
-        return Web3.Utils.publicToAddress(publicKey)
+        return Web3Utils.publicToAddress(publicKey)
     }
     
     static public func hashECRecover(hash: Data, signature: Data) -> EthereumAddress? {
@@ -229,7 +229,7 @@ extension Web3.Utils {
         let vData = signature[64]
         guard let signatureData = SECP256K1.marshalSignature(v: vData, r: rData, s: sData) else {return nil}
         guard let publicKey = SECP256K1.recoverPublicKey(hash: hash, signature: signatureData) else {return nil}
-        return Web3.Utils.publicToAddress(publicKey)
+        return Web3Utils.publicToAddress(publicKey)
     }
     
     /// returns Ethereum variant of sha3 (keccak256) of data. Returns nil is data is empty