+++ Example/web3swiftExample/web3swiftExample/ViewController.swift
@@ -123,7 +123,7 @@ class ViewController: UIViewController {
         var tokenTransferOptions = Web3Options.defaultOptions()
         tokenTransferOptions.gasPrice = gasPriceRinkeby
         tokenTransferOptions.from = ks?.addresses?.first!
-        let testToken = web3Rinkeby.contract(Web3.Utils.erc20ABI, at: EthereumAddress("0xa407dd0cbc9f9d20cdbd557686625e586c85b20a")!, abiVersion: 2)!
+        let testToken = web3Rinkeby.contract(Web3Utils.erc20ABI, at: EthereumAddress("0xa407dd0cbc9f9d20cdbd557686625e586c85b20a")!, abiVersion: 2)!
         let intermediateForTokenTransfer = testToken.method("transfer", parameters: [EthereumAddress("0x6394b37Cf80A7358b38068f0CA4760ad49983a1B")!, BigUInt(1)] as [AnyObject], options: tokenTransferOptions)!
         let gasEstimateResult = intermediateForTokenTransfer.estimateGas(options: nil)
         guard case .success(let gasEstimate) = gasEstimateResult else {return}