+++ web3swiftTests/RinkebyPersonalSignatureTests.swift
@@ -27,7 +27,7 @@ class RinkebyPersonalSignatureTests: XCTestCase {
         print("V = " + String(unmarshalledSignature.v))
         print("R = " + Data(unmarshalledSignature.r).toHexString())
         print("S = " + Data(unmarshalledSignature.s).toHexString())
-        try! print("Personal hash = " + Web3.Utils.hashPersonalMessage(message).toHexString())
+        try! print("Personal hash = " + Web3Utils.hashPersonalMessage(message).toHexString())
         let signer = try web3.personal.ecrecover(personalMessage: message, signature: signature)
         XCTAssert(expectedAddress == signer, "Failed to sign personal message")
     }
@@ -46,7 +46,7 @@ class RinkebyPersonalSignatureTests: XCTestCase {
         print("V = " + String(unmarshalledSignature.v))
         print("R = " + Data(unmarshalledSignature.r).toHexString())
         print("S = " + Data(unmarshalledSignature.s).toHexString())
-        try print("Personal hash = " + Web3.Utils.hashPersonalMessage(messageData).toHexString())
+        try print("Personal hash = " + Web3Utils.hashPersonalMessage(messageData).toHexString())
         let jsonString = "[{\"inputs\":[],\"payable\":false,\"stateMutability\":\"nonpayable\",\"type\":\"constructor\"},{\"constant\":true,\"inputs\":[{\"name\":\"_message\",\"type\":\"string\"}],\"name\":\"hashPersonalMessage\",\"outputs\":[{\"name\":\"hash\",\"type\":\"bytes32\"}],\"payable\":false,\"stateMutability\":\"pure\",\"type\":\"function\"},{\"constant\":true,\"inputs\":[{\"name\":\"_message\",\"type\":\"string\"},{\"name\":\"v\",\"type\":\"uint8\"},{\"name\":\"r\",\"type\":\"bytes32\"},{\"name\":\"s\",\"type\":\"bytes32\"}],\"name\":\"recoverSigner\",\"outputs\":[{\"name\":\"signer\",\"type\":\"address\"}],\"payable\":false,\"stateMutability\":\"pure\",\"type\":\"function\"}]"
         let contract = try web3.contract(jsonString, at: "0x6f1745a39059268e8e4572e97897b50e4aab62a8")
         var options = Web3Options.default
@@ -54,7 +54,7 @@ class RinkebyPersonalSignatureTests: XCTestCase {
         var intermediate = try contract.method("hashPersonalMessage", args: message, options: options)
         var res = try intermediate.call(options: nil)
         guard let hash = res["hash"]! as? Data else { return XCTFail() }
-        let hash2 = try Web3.Utils.hashPersonalMessage(messageData)
+        let hash2 = try Web3Utils.hashPersonalMessage(messageData)
         XCTAssert(hash2 == hash)
 
         intermediate = try contract.method("recoverSigner", args: message, unmarshalledSignature.v, Data(unmarshalledSignature.r), Data(unmarshalledSignature.s), options: options)